requirements.txt
unicorn>=2.1.1,<3.0.0
capstone>=3.0.4,<4.0.0
pycryptodome>=3.21.0